fix: adequa swagger ao padrao generation

+++ src/main/java/com/generation/projetorh/controller/DepartamentoController.java
@@ -62,7 +62,13 @@ public class DepartamentoController {
 	}
 
 	@DeleteMapping("/{id}")
-	public void delete(@PathVariable Long id) {
-		departamentoRepository.deleteById(id);
+	public ResponseEntity<Void> delete(@PathVariable Long id) {
+
+		if (departamentoRepository.existsById(id)) {
+			departamentoRepository.deleteById(id);
+			return ResponseEntity.noContent().build();
+		}
+
+		return ResponseEntity.notFound().build();
 	}
 }
